Extract SharedPRNG class
